Created the setImage method in each piece subclass & defined a checkLegalMoves
checkLegalMoves was left with nothing in it for now, and we are trying to find a better way to utilize the setImage class, since the current process feels very cluttered and repetitive to the point that it warrants a spot in the parent class.  We just need to figure out how to do that.